Reduce probability of deadlocks in opaque table registration.
This uses the operation that has been intentionally hardened against
deadlocks (sync) to reduce the chance that we try an operation I don't
know how to harden in general (ensureTableExists).

But I don't think it eliminates the possibility altogether, because
we've actually observed a deadlock with one participant inside
getExistingTable, not just ensureTableExists - that shouldn't happen
because the former should be a read-only operation that doesn't ever
need to acquire an exclusive lock, but we can't control what the
sqlalchemy inspection code does.