Rework handling of governor dimension restrictions.
It took me a while to figure out why the set-like
GovernorDimensionRestriction class kept needing new set-op method
variants for ~every context in which it was being used, but the answer
is that it was really a mapping of sets (one for each governor
dimension), and hence when instances have different dimensions
(mapping keys) dimensions, their relationship is not set-like.

The best solution seems to be to remove that class in favor of the
kind of mapping-of-sets that backed it, and move its convenience
functionality that's important for CollectionSummary to
CollectionSummary itself.  In some cases we now use the new
UnboundableSet class from lsst.utils, which lets us represent the full
set in abstract without being able to enumerate all of the possible
values.

I changed some attribute names and the module name for collecton
summary to get them consistent with now-preferred style; it made sense
to do this at the same time because I wanted to find all code using
CollectionSummary in order to make sure it was updated appropriately,
and renames made MyPy very good at that.  I also switched from
Dimension instance keys to `str` dimension name keys, since I think
we'll be doing more of that with RFC-834 adopted.

Finally, I switched the governor dimension restriction logic in the
query expression analysis code to track all dimensions, not just
governors, since it was actually more effort to pay attention to the
governors only and the bounds on the others might prove useful for
query optimization in the future.